after ICIS 2023:
--add the 3D surface for robustness across p1, p2
--add interaction: threshold + inactive; threshold + asymmetry; threshold + incentive
--remove parameter s
--add the Simplified Version to reproduce the main results of criticality (threshold, asymmetry, incentive)
--regularize the code norm including parameter type and default value
--add Random Network Topology for Robustness -> Become a justification for the isolation structure

TODO: to measure how the incentive will bias the token distribution (new comer? or lucky comer) -> This is more like a thought experiment -> Compared to ideally equal decision power distribution, overly used incentive will centralize the decision power?

+++ V4/Incentive/dao_run_2.py
@@ -15,7 +15,7 @@ import pickle
 import math
 
 
-def func(m=None, n=None, group_size=None, lr=None, incentive=None, inactive_rate=None,
+def func(m=None, n=None, group_size=None, lr=None, incentive=None, active_rate=None,
          search_loop=None, loop=None, return_dict=None, sema=None):
     np.random.seed(None)
     reality = Reality(m=m)
@@ -29,25 +29,25 @@ def func(m=None, n=None, group_size=None, lr=None, incentive=None, inactive_rate
             dao.search(threshold_ratio=0.5)
     else:
         for _ in range(search_loop):
-            dao.incentive_search(threshold_ratio=0.5, incentive=incentive, inactive_rate=inactive_rate)
-    return_dict[loop] = [dao.performance_across_time, dao.consensus_performance_across_time,
-                         dao.diversity_across_time, dao.variance_across_time, dao.gini_across_time]
+            dao.incentive_search(threshold_ratio=0.5, incentive=incentive, active_rate=active_rate)
+    return_dict[loop] = [dao.performance_across_time, dao.consensus_performance_across_time, dao.diversity_across_time, dao.variance_across_time]
     sema.release()
 
 
 if __name__ == '__main__':
+    import os
     t0 = time.time()
     m = 90
     n = 350
     lr = 0.3
-    repetition = 200
+    repetition = 50
     incentive_list = [0.3, 0.4]
-    inactive_rate_list = [0, 0.1, 0.2, 0.3, 0.4]
+    active_rate_list = [0.9, 0.8, 0.7, 0.6]
     search_loop = 500
     group_size = 7  # the smallest group size in Fang's model: 7
     concurrency = 50
     for incentive in incentive_list:
-        for inactive_rate in inactive_rate_list:
+        for active_rate in active_rate_list:
             sema = Semaphore(concurrency)
             manager = mp.Manager()
             return_dict = manager.dict()
@@ -55,7 +55,7 @@ if __name__ == '__main__':
             for loop in range(repetition):
                 sema.acquire()
                 p = mp.Process(target=func,
-                               args=(m, n, group_size, lr, incentive, inactive_rate, search_loop, loop, return_dict, sema))
+                               args=(m, n, group_size, lr, incentive, active_rate, search_loop, loop, return_dict, sema))
                 jobs.append(p)
                 p.start()
             for proc in jobs:
@@ -65,46 +65,56 @@ if __name__ == '__main__':
             consensus_across_repeat = [result[1] for result in results]
             diversity_across_repeat = [result[2] for result in results]
             variance_across_repeat = [result[3] for result in results]
-            gini_across_repeat = [result[4] for result in results]
             # After taking an average across repetitions
             performance_final = []
             consensus_final = []
             diversity_final = []
             variance_final = []
-            gini_final = []
             for period in range(search_loop):
                 performance_temp = [performance_list[period] for performance_list in performance_across_repeat]
                 consensus_temp = [consensus_list[period] for consensus_list in consensus_across_repeat]
                 diversity_temp = [diversity_list[period] for diversity_list in diversity_across_repeat]
                 variance_temp = [variance_list[period] for variance_list in variance_across_repeat]
-                gini_temp = [gini_list[period] for gini_list in gini_across_repeat]
 
                 performance_final.append(sum(performance_temp) / len(performance_temp))
                 consensus_final.append(sum(consensus_temp) / len(consensus_temp))
                 diversity_final.append(sum(diversity_temp) / len(diversity_temp))
                 variance_final.append(sum(variance_temp) / len(variance_temp))
-                gini_final.append(sum(gini_temp) / len(gini_temp))
 
-            with open("dao_performance_incentive_{0}_inactive_{1}".format(incentive, inactive_rate), 'wb') as out_file:
-                pickle.dump(performance_final, out_file)
-            with open("dao_consensus_performance_incentive_{0}_inactive_{1}".format(incentive, inactive_rate), 'wb') as out_file:
-                pickle.dump(consensus_final, out_file)
-            with open("dao_diversity_incentive_{0}_inactive_{1}".format(incentive, inactive_rate), 'wb') as out_file:
-                pickle.dump(diversity_final, out_file)
-            with open("dao_variance_incentive_{0}_inactive_{1}".format(incentive, inactive_rate), 'wb') as out_file:
-                pickle.dump(variance_final, out_file)
-            with open("dao_gini_incentive_{0}_inactive_{1}".format(incentive, inactive_rate), 'wb') as out_file:
-                pickle.dump(gini_final, out_file)
+            performance_file_name = r"dao_performance_incentive_{0}_active_{1}".format(incentive, active_rate)
+            delay = np.random.uniform(1, 20)
+            time.sleep(delay)
+            if os.path.exists(performance_file_name):
+                with open("dao_performance_incentive_{0}_inactive_{1}".format(incentive, active_rate), 'rb') as infile:
+                    prior_performance = pickle.load(infile)
+                with open("dao_consensus_performance_incentive_{0}_inactive_{1}".format(incentive, active_rate), 'wb') as infile:
+                    prior_consensus = pickle.load(infile)
+                with open("dao_diversity_incentive_{0}_inactive_{1}".format(incentive, active_rate), 'wb') as infile:
+                    prior_diversity = pickle.load(infile)
+                with open("dao_variance_incentive_{0}_inactive_{1}".format(incentive, active_rate), 'wb') as infile:
+                    prior_variance = pickle.load(infile)
+                performance_final = [(each_1 + each_2) / 2 for each_1, each_2 in zip(prior_performance, performance_final)]
+                consensus_final = [(each_1 + each_2) / 2 for each_1, each_2 in zip(prior_consensus, consensus_final)]
+                diversity_final = [(each_1 + each_2) / 2 for each_1, each_2 in zip(prior_diversity, diversity_final)]
+                variance_final = [(each_1 + each_2) / 2 for each_1, each_2 in zip(prior_variance, variance_final)]
+                with open("dao_performance_incentive_{0}_inactive_{1}".format(incentive, active_rate), 'wb') as out_file:
+                    pickle.dump(performance_final, out_file)
+                with open("dao_consensus_performance_incentive_{0}_inactive_{1}".format(incentive, active_rate), 'wb') as out_file:
+                    pickle.dump(consensus_final, out_file)
+                with open("dao_diversity_incentive_{0}_inactive_{1}".format(incentive, active_rate), 'wb') as out_file:
+                    pickle.dump(diversity_final, out_file)
+                with open("dao_variance_incentive_{0}_inactive_{1}".format(incentive, active_rate), 'wb') as out_file:
+                    pickle.dump(variance_final, out_file)
 
-            with open("dao_original_performance_incentive_{0}_inactive_{1}".format(incentive, inactive_rate), 'wb') as out_file:
-                pickle.dump(performance_across_repeat, out_file)
-            with open("dao_original_consensus_performance_incentive_{0}_inactive_{1}".format(incentive, inactive_rate), 'wb') as out_file:
-                pickle.dump(consensus_across_repeat, out_file)
-            with open("dao_original_diversity_incentive_{0}_inactive_{1}".format(incentive, inactive_rate), 'wb') as out_file:
-                pickle.dump(diversity_across_repeat, out_file)
-            with open("dao_original_variance_incentive_{0}_inactive_{1}".format(incentive, inactive_rate), 'wb') as out_file:
-                pickle.dump(variance_across_repeat, out_file)
-            with open("dao_original_gini_incentive_{0}_inactive_{1}".format(incentive, inactive_rate), 'wb') as out_file:
-                pickle.dump(gini_across_repeat, out_file)
+            else:
+                with open("dao_performance_incentive_{0}_inactive_{1}".format(incentive, active_rate), 'wb') as out_file:
+                    pickle.dump(performance_final, out_file)
+                with open("dao_consensus_performance_incentive_{0}_inactive_{1}".format(incentive, active_rate), 'wb') as out_file:
+                    pickle.dump(consensus_final, out_file)
+                with open("dao_diversity_incentive_{0}_inactive_{1}".format(incentive, active_rate), 'wb') as out_file:
+                    pickle.dump(diversity_final, out_file)
+                with open("dao_variance_incentive_{0}_inactive_{1}".format(incentive, active_rate), 'wb') as out_file:
+                    pickle.dump(variance_final, out_file)
+
     t1 = time.time()
     print(time.strftime("%H:%M:%S", time.gmtime(t1 - t0)))
